+++ controller/lib/observer/__init__.py
@@ -1,0 +1,3 @@
+from .classes import Subject, InfoViewer
+
+__all__ = ["Subject","InfoViewer"]
